split navigation in more components
